requirements.txt
# Core Django
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
asgiref==3.11.0
sqlparse==0.5.5

# Server
gunicorn==23.0.0
whitenoise==6.7.0

# Database
psycopg2-binary==2.9.11

# Media (optional, you used earlier)
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
pillow==12.1.0

# HTTP
requests==2.32.5
urllib3==2.6.3
charset-normalizer==3.4.4
certifi==2026.1.4
idna==3.11
six==1.17.0
packaging==25.0

# ===== ML / DL (CRITICAL) =====
numpy==2.0.0
pandas==2.2.2
scikit-learn==1.5.1
torch==2.2.2
sentence-transformers==2.7.0
